Ensure entire node points to allocated memory
Adapted from https://github.com/cockroachdb/pebble/pull/669. Only a
few small changes were needed to, presumably because pebble's fork
of this library has diverged slightly.

Previously, if a node was allocated at the end of the arena, the node
struct could straddle the end of the arena, with the unused links
extending into unallocated memory.

Go 1.14 adds a checkptr check to -race builds that checks for pointer
alignment issues, and it flags the conversion of these
partially-allocated nodes at the arena boundary.

This commit changes the allocation logic to grow the arena such that the
entirety of the last node struct is allocated.